Revert "Some updates to elastix.m. Due to a current limit to function of elastix, multiple-channel registration results in only one channel output. Transformix is used to transform other channels." The purpose of elastix.m is to register two images in the sense of computing the transform that aligns them. The main output is the transform, the deformed image produced by the elastix binary is just a side product.
If one wants to then apply the transform to the moving image, that's what transformix.m is for. If the image is RGB, transformix.m will produce an RGB image.

The reason for this is that applying transformix adds substantial computational cost. This may not be desired. For example, we work aligning stacks of histology images. We may need several sweeps of the stack registering each slice to its neighbours. But we don't really need to produce the colour deformed images after every sweep. It's faster to first compute a transform in every sweep, and then compose the transforms and apply them all together to the initial images.
This reverts commit 52daf56141678e3bf1dc876a6a1e13402321a0c2.