Exclude deleted .docx files from pandoc conversion
If you try to commit changes to your repo that include the deletion of a tracked *.docx file, pandoc will try to convert the (now nonexistent) file to *.md, and fail. Since the script executed by the pre-commit hook therefore returns an error, the commit will fail.

A simple workaround is to use the `--diff-filter=d` option, which will *exclude* any deleted files from being passed to pandoc. This creates a need to remove the associated *.md file, which I have added as a separate `for` loop at the bottom.